+++ tests/integration_tests/functional/test_fuzzing.py
@@ -3,6 +3,7 @@
 """A test that ensures the fuzzing feature warning appears at startup."""
 
 import host_tools.cargo_build
+from framework.artifacts import GUEST_KERNEL_DEFAULT, pin_guest_kernel
 from framework.microvm import MicroVMFactory
 
 
@@ -25,13 +26,14 @@ def test_fuzzing_warning(guest_kernel, rootfs):
     ), f"Version string missing +fuzzing suffix:\n{uvm.log_data}"
 
 
-def test_no_fuzzing_warning(uvm_plain):
+@pin_guest_kernel(GUEST_KERNEL_DEFAULT)
+def test_no_fuzzing_warning(uvm):
     """Checks that a standard Firecracker binary does not log the fuzzing warning"""
 
-    uvm_plain.spawn()
-    uvm_plain.basic_config()
-    uvm_plain.start()
+    uvm.spawn()
+    uvm.basic_config()
+    uvm.start()
 
     assert (
-        "built with the `fuzzing` feature enabled" not in uvm_plain.log_data
-    ), f"Fuzzing warning unexpectedly found in logs:\n{uvm_plain.log_data}"
+        "built with the `fuzzing` feature enabled" not in uvm.log_data
+    ), f"Fuzzing warning unexpectedly found in logs:\n{uvm.log_data}"