+++ tests/integration_tests/functional/test_gdb.py
@@ -12,6 +12,7 @@ from pathlib import Path
 import pytest
 
 import host_tools.cargo_build
+from framework.artifacts import GUEST_KERNEL_DEFAULT_DEBUG, pin_guest_kernel
 from framework.microvm import MicroVMFactory
 
 
@@ -19,13 +20,14 @@ from framework.microvm import MicroVMFactory
     platform.machine() != "x86_64",
     reason="GDB requires a vmlinux but we ship a uImage for ARM in our CI",
 )
-def test_gdb_connects(guest_kernel_default_debug, rootfs):
+@pin_guest_kernel(GUEST_KERNEL_DEFAULT_DEBUG)
+def test_gdb_connects(guest_kernel, rootfs):
     """Checks that GDB works in a FC VM"""
 
     bin_dir = host_tools.cargo_build.build_gdb()
 
     vmfcty = MicroVMFactory(bin_dir)
-    uvm = vmfcty.build(guest_kernel_default_debug, rootfs)
+    uvm = vmfcty.build(guest_kernel, rootfs)
     uvm.spawn(validate_api=False)
     uvm.add_net_iface()
     uvm.basic_config()
@@ -54,7 +56,7 @@ def test_gdb_connects(guest_kernel_default_debug, rootfs):
             echo 'waiting for {chroot_gdb_socket}';
             sleep 1;
         done;
-        gdb {guest_kernel_default_debug} -batch -x {gdb_script}
+        gdb {guest_kernel} -batch -x {gdb_script}
         """,
         shell=True,
         stdout=subprocess.PIPE,